feat(firewall): Drop undesired external network traffic
The firewall is set-up to drop all traffic to a public network interface
at the end of filter chain named CLUSTER.  Both DOCKER_USER and INPUT
chains jump to CLUSTER.  Ports can be made accessible to the public by
setting the inventory var externally_accessible_network_ports.  This
set-up won't intefere with Docker's use of iptables, as long as Docker
is not listening on the public interface.  This means nodes should
advertise the IP address of a private interface.  Set the inventory var
log_dropped_network_traffic=true to log at level INFO.