feat(baseline): Memory management sysctl overrides
When the var "discourage_swapping" is set (and true), set
vm.swappiness=0 to avoid swapping as much as possible (for machines with
abundant physical memory) and set vm.overcommit_memory=1 to allow memory
allocations to succeed even when it looks like physical memory is
oversubscribed.  These overrides are especially useful for hosts running
MariaDB and Redis.