feat: Stackmanager will load stacks from a var
in addition to loading stacks from a manifest file.  Define a var named
"stacks" (and make sure no manifest file is present) and the
stackmanager role will deploy them in the usual way.  An additional
stack var, "path", is understood to be the name of the local (ansible
control host) dir holding the stack's files and may be different from
the stack name.  This allows a single directory of stack dirs, avoiding
collisions of stack name.  For example there could be stacks/foo-traefik
and stack/bar-traefik stacks, both named "traefik".