feat(stackmanager): Rudimentary support for multi-tenant stacks
Stack manager unserstands a new stack var "tenant_var" which points to
an ansible variable that holds a list of the stack's tenants.

Stack manager processes a stack's tenants so that the stack can generate
a docker config for each tenant, e.g. a symfony dotenv file.

Stack manager handles .j2mt stack files which are treated like ordinary
.j2 templates, but are processed once per-tenant.

A tenant is expected to look like this:

    - id: blue
      domain: nebulous-blue.local
      stackfiles:
        - template: .env.j2mt
          docker_config:
            alias: env-blue
            file: .env.blue # filename in the remote stack dir ...
            target: nebulous-blue.local-env # .. in the container