feat(app): Create Postgres databases and database users
Automate the creation of postgres databases and database users.  For any
node hosting a postgres service, set the inventory variables "databases"
and "database_users".  Set the variable "pg_database_admin" to an
account which can create databases and users.

An example of inventory:

mygroup:
  hosts:
    db1:
      database_admin:
        user: root
        port: 33301
        password: '{{ db1_MariadbPrimaryRootPassword }}'
      pg_database_admin:
        user: postgres
        port: 54321
        password: '{{ db1_PostgresSuperuserPassword }}'
      databases:
        - name: mydb
        - name: totherdb
          family: pg
      database_users:
        - name: myuser
          password: '{{ db1_MyuserPassword }}'
          privileges: '*.*:ALL,GRANT'
          resource_limits:
            MAX_USER_CONNECTIONS: 3
        - name: totheruser
          password: '{{ db1_TotheruserPassword }}'
          privileges: CREATE,USAGE
          db: totherdb

Ansible will create Postgres databases for those which have `family=pg` and
Postgres users for those which have a `db` property (and the user will own the
given db).