feat(baseline): Disable password and root SSH auth
The network is hostile and hosts must expect it to continually attempt
brute-force password authentication with our hosts.  This change means
they're unlikely to succeed.

CAUTION!  This config is not applied if you're running ansible as root!
	  This avoids denying yourself the ability to complete the play.
It's common to run ansible as root against a freshly provisioned host.
Ideally, only the basline layer needs to be applied as root, then you
can run ansible as your own user (and re-run baseline.config).

Avoid needing to use confold with apt upgrade by setting the ssh config
in an included file.